Configure suricata.yaml
Remove af-packet interface [1] because it does not exist and causes suricata to fail to start. - This might be a problem in environments where 2 NICs are configured - but I'll cross that bridge when I come to it.

Replace af-packet interface [0] dynamically based on primary NIC

Replace pfring interface [0] dynamically based on primary NIC. Not sure if this is required, but done for completeness.